+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Added
 
+- Added `apps` command that shows a list of available apps.
 - Added a warning when multiple versions are found.
 - Added colorful output to the message stating if there was no version found.